ensured long datatype for global ID endpoints

+++ eRecruiter.Api.Client/Requests/Agnostic/AgnosticApplicantsRequest.cs
@@ -6,7 +6,7 @@ namespace eRecruiter.Api.Client.Requests
 {
     public class AgnosticApplicantsByGlobalIdRequest : HttpRequestMessage<IEnumerable<AgnosticApplicantsResponse>>
     {
-        public AgnosticApplicantsByGlobalIdRequest(int globalId)
+        public AgnosticApplicantsByGlobalIdRequest(long globalId)
             : base(HttpMethod.Get, "/Api/Agnostic/Applicants/ByGlobalId/" + globalId)
         {
         }
